New synchronize function
I just realized that there appears to be a bug: the functions should
be wrapped on a fiber for them to work, but they work regardless.
This is likely because there’s some leftover fiber from the futures tests,
for reasons I don’t know.  If the tests for futures don’t execute, the
synchronize tests fail like they should.

Committing so I can push it while I review.